Fork of version 0.3 to version 1.0.
The SVN repository is now up to date with all compressed tar files having been imported into the
repository.  The repository consists of the compressed tar files of the initial 'modelfree' program,
the rename to 'mf', and the final program 'relax'.  The date on every revision reflects that of the
compressed tar file it corresponds to.

Development and proper release versioning using this repository can now commence.