Change of the method for converting between internal parameter representation and external representation.
Two specific function types have been added.  The function 'return_convesion_factor' takes the name
of the parameter as an argument and returns the factor of conversion between internal and external
representations.  The function 'return_units' takes the name of the parameter as an argument and
returns a string representation of the units, for example 'Angstrom'.

The results reading and writing functions in 'specific_fns/model_free.py' were modified to use these
functions.