Bug fix:
Fixed a bug which manifested itself in the full analysis sample script when optimising the diffusion
tensor after reading in the local tm results.  The bug was in the add_residue function of the
relaxation data specific functions class and was because the 'data.noe_r1_table' data structure was
not being properly created.