Massive bug fix!!!
Caught the bug causing problems with Newton minimisation of the diffusion parameters together with
the model-free parameters.  The scaling matrix was not being constructed properly!  Only the first
residue's parameters were being scaled 'index = 0', all the rest were unscaled.  The bug was because
the model-free specific minimise function was passing in the index of zero rather than None!!!