Bug fixes.
All three Euler angles are now properly wrapped between 0 and 2pi.

The glide reflection symmetry of the Euler angle beta (0 to pi and pi to 2pi) is now taken into
account.

The glide reflection and translation symmetries are now removed, hence the Euler angles will all be
between 0 and pi.