Further modification to the packaging script and added the program version to the manual.
The checked out repository is checked for changes before packaging.

A script has been created to make the file 'docs/latex/relax_version.tex', the contents of which are
included in the cover of the manual.

In the packaging script, the manual version is checked against the program version to see if they
match.