Creating an unstable development fork from 1.0 called 1.1.
This fork exists because the last few commits of the relaxation curve fitting modifications have
involved changes which significantly alter parts of the program and could cause the program not to
work in certain situations.  Therefore the stable branch 1.0 has been forked and the unstable 1.1
branch has been created for developing the relaxation curve fitting code.

One of the major changes includes the addition of C code for the implementation of functions
involved in the optimisation of the parameters of the exponential curves.  This code is compiled
into shared objects '*.so' which are loaded at run-time as modules into the Python code.  The
interface between C and Python is the low level Python/C API.

To build these C modules, relax has shifted to using Makefiles.  Therefore to compile the code, the
command 'make' will create the shared objects.  Instead of relying on a script, 'make clean' will
remove all the '*.pyc', '*.o', and '*.so' files.  The Makefile will be extended to include
procedures which include the creation of the relax manual, creation of the HTML version of the
manual, and the compilation and packaging of relax for distribution.

C code has been introduced solely to decrease computation time.  This inclusion also allows the use
of fast Fortran linear algebra libraries using C to Fortran interfaces.  Despite using C code, relax
will be developed so that the compilation of this code is not essential for running the program.  If
the shared objects do not exist, relax will disable these components of the program.

The 1.0 branch will be reverted back to revision r2294 which was the last commit not involving
relaxation curve fitting.  This stable branch is now reserved for bug fixes only, features will be
added to the 1.1 branch.