Relaxation curve fitting and Makefile changes.
A second Makefile within the 'maths_fns' directory has been created as a copy of the Makefile in the
base directory.  The new Makefile contains the compilation rules for the maths_fns directory which
were located in the old file.

The files 'exp_fn.c' and 'exp_fn.h' have been renamed to 'relax_fit.c' and 'relax_fit.h'
respectively.  This is so the old function names can contain the code for back calculating the
peak intensities from the given parameters, while the relax_fit code contains the setup function as
well as the functions func, dfunc, and d2func.

More minimisation code has been added to 'specific_fns/relax_fit.py'.