A number of changes to the code for Art Palmer's Modelfree4 program.
The 'palmer.create()' and 'palmer.execute()' now have the keyword argument 'binary'.  If the default
of 'modelfree4' is changed to another name, then the binary file executed will be different.  This
is useful for running a different version or a binary which is not in the environment's path.

The precision of the data in the 'mfdata' file has been significantly improved.

The function 'palmer.extract()' has been rewritten to be much more efficient.  Rather than walking
through the 'mfout' file line by line for each parameter for each residue, only one parse of the
file is done to work out the position of each section.  Then the section position plus the residue
position is added to extract the relevant data.