Bug fixes due to unstable, broken Numeric Python packages:
Fixed a bug in the minimisation code.  The Cholesky Hessian modification was failing due to the
Numeric Python package instability.  The code should catch if the Cholesky factorisation fails and
then modify the factor tau and repeat the steps until the factorisation is succesful.  Previously
the Numeric 'cholesky_decomposition()' function would throw out the "LinearAlgebraError" error, but
now something else is thrown out!  The code now handles any error.

A bug fix to the Newton minimisation code has also been added.  When the print flag is set above 3,
the code tries to calculate the eigenvalues of the Hessian.  However the Numeric function
'eigenvectors()' is completely broken and causes any code which calls it to hang infinitely.  The
code has been commented out.