Major changes to the development chapter of the manual.
The relax-announce mailing list is now included in the intro.

All sections have been significantly edited.

An empty subsection called 'Variable, function, and class names' has been added.

The section 'Committing conventions' has been added and the subsection 'Format of the commit logs'
has been written.

The 'Making changes' section has been added and consists of two subsections, 'Patching' which came
from the start of the chapter and 'Becoming a committer' which describes what is needed to become a
committer.