The creation of a framework for the generation of Molmol macros.
A large number of changes and additions have been made including:

The user function 'molmol.write()' has been created.

A number of functions in the generic molmol code have been renamed.  These are 'open_pipe' to
'pipe_open', 'pipe_open' to 'pipe_open_test', and 'write' to 'pipe_write'.

The functions 'self.write()' and 'self.create_macro()' have been added to 'generic_fns/molmol.py'.
The first function calls the second which returns an array of Molmol commands.  The point of
splitting the functions is so in the future 'self.exec_macro()' can be added to pipe the command
directly into Molmol.

The two functions 'self.molmol_macro()' and 'self.molmol_macro_classic()' have been added to the
model-free specific code.  The second function currently returns a dummy array so that the resultant
macro is useless.  The first function is returned by the specific_setup.py code.

A new error class called 'RelaxStyleError' has been created.  This error is to be called anytime an
argument called 'style' is unrecognised.