Completion of task #3146, mapping parameters onto structure through Molmol macros.
The task is located as https://gna.org/task/?func=detailitem&item_id=3146.

Arbitrary colour gradients are supported.  To allow this, the 'colour_start' and 'colour_end'
keyword arguments have been added to the 'molmol.write()' user function.  These can either be colour
names like 'red' or an RGB (red, green, blue) array such as [1.0, 1.0, 1.0] which is white.  The
file colour.py has been added to the base directory for colour support.  It contains two functions,
the first being 'self.linear_gradient()' which takes a value between zero and one and determines the
RGB vector of that value when zero is the start colour and one is the end colour.  The second
function, 'self.rgb()' takes a string representing a colour and translates it to an RGB array.

For the 'classic' style, only the mapping of model-free parameters is currently supported.  The
function 'self.molmol_macro_classic()' has been added to the model-free specific code.  A number of
auxiliary functions 'self.molmol_macro_classic_*()' are called by this main function to generate the
Molmol commands for colouring, setting the width, and setting the style of backbone bonds, for
creating the Molmol macro header, and for translating order parameters, correlation times, and
chemical exchange into values between zero and one for the colour gradients.

A number of new RelaxError objects have been created for the colours, etc.