A number of large additions to the relax manual.
In the development chapter a section called the 'The core design of relax' has been added.  This
section has a diagram of the internal design of relax and text describing all the parts.

The 'Values, gradients, and Hessians' chapter has been completed.  All the first and second partial
derivatives of the weights and correlation times of the ellipsoid, spheroid, and sphere have been
added (as well as the original equations).  The equations, gradients, and Hessians of the dot
products used in the weights of the ellipsoid and spheroid have been added.  Importantly, the
model-free equations and the first and second partial derivatives which create the gradients and
Hessians have been added.  These equations include the derivatives of the original model-free
formula of Lipari and Szabo (1982a,b) with the parameters {S2, te}, the derivatives of the extended
model-free formula of Clore (1990) with the parameters {S2, S2f, tf, ts}, and the derivatives of an
alternative set of extended parameters {S2f, S2s, tf, ts}.