Modifications to the notation used in the generic model-free equations, gradients, and Hessians.
The geometric parameters {Diso, Da, Dr} of the ellipsoid, {Diso, Da} of the spheroid, and {Diso} of
the sphere have been changed from the Dj and Dk notation to the new set theory notation Gj and Gk.
This is because the geometric parameter set is denoted by the symbol G in Gothic font.

The summation notation has also been changed.  Instead of the summation index i ranging from m to n
it now ranges from -k to k.