Finally, a real fix to bug #5746.
The bug report is located at https://gna.org/bugs/?func=detailitem&item_id=5746 and is discussed in
the thread starting at https://mail.gna.org/public/relax-devel/2006-04/msg00018.html.

The fix follows from an idea in the post located at
https://mail.gna.org/public/relax-devel/2006-04/msg00021.html.  Firstly the model-free calculate()
function was modified so that when self.param_set is equal to 'all' or 'diff' (global statistics)
then the chi-squared value calculated is summed and placed in the global data structure
self.relax.data.chi2 rather than the residue specific data structure.

This change allowed the 'combined' flag to be eliminated from the self.model_statistics() function
of 'specific_fns/model_free.py'.  This flag was the source of all the problems and it's removal
allowed chunks of obscure code to be axed.  The statistics only need to refer to the self.param_set
value to determine which stats should be returned.