Fix to bug #6403.
The bug is located at https://gna.org/bugs/?func=detailitem&item_id=6403.

The problem is that relax will not start due to an ImportError when trying to import the object
'devnull'.  The solution was to import the object using a try statement to allow execution of the
program.  However if relax ever attempts to use 'devnull' a RelaxError is thrown, a message saying
to upgrade to Python >= 2.4 is printed, and the program terminates.  This is currently only a
problem in the test-suite.