relax version 1.2.6.
A quick release to fix the fatal 'devnull' ImportError (bug #6403).  This bug effects those using
Python versions < 2.4 and prevents relax from starting up.  The solution was to import the object
using a try statement to allow execution of the program.  However if relax ever attempts to use
'devnull' a RelaxError is thrown, a message saying to upgrade to Python >= 2.4 is printed, and the
program terminates.  This is currently only a problem in the test-suite.

The full list of changes is:

    Features:
        N/A

    Changes:
        N/A

    Bugfixes:
        * Added the missing PDB file used by one of the OpenDX tests of the relax test-suite. (bug
        #6404)
        * Fixed the fatal devnull import failure for Python versions < 2.4. (bug #6403)