A warning system for relax.
This is based on the system started in r2532, and is discussed in 
detail in the thread starting:
https://mail.gna.org/public/relax-devel/2006-08/msg00071.html

The system is based on the standard Python warnings system.
Warnings are generated with the syntax:

warn(RelaxWarning(args))

where RelaxWarning can be any of the Warning classes defined in
errors.py, and args is an arbitrary number of arguments of 
arbitrary type which are used by the Warning class to generate a
meaningful message string.

The system also impliments a --pedantic command-line flag, which
causes all warnings to be escalated to exceptions. All of the 
relax warnings are subclasses of both BaseWarning and BaseError,
so the escalated exceptions will be caught by except clauses
based on these types.

A further feature in the current implimentation is to include
an exception-like traceback in the warning message when in
debug mode.