Fix to bug #7225.
This bug was reported by Alex Hansen at https://gna.org/bugs/index.php?7225.

To enable comprehensible error messages when attempting to execute Modelfree4, the function
'self.test_binary()' was added to the file 'io.py'.  This was described in the post located at
https://mail.gna.org/public/relax-devel/2006-09/msg00034.html (Message-id:
<7f080ed10609280917s139d2c2cm9f0edcbd5a9fc04c@mail.gmail.com>).  The implementation is slightly
different to the post though.  Firstly the supplied string corresponding to the binary is tested to
see if the full path has been supplied.  If so then the string is tested to see if the file exists
and if the file is executable.  Otherwise the file is searched for in the PATH environmental
variable.  Point 5 of the post at https://mail.gna.org/public/relax-devel/2006-09/msg00034.html,
testing that the binary is in the current directory was not implemented.  The reason is because I
was unable to successfully modify the PATH variable.

Three new RelaxErrors have been created to print out comprehensible messages:
RelaxMissingBinaryError, RelaxNonExecError, and RelaxNoInPathError.