Fix for bug #7226 (https://gna.org/bugs/?7226).
This bug was reported by Alex Hansen.  The problem was caused by the model-free results file reading
function, specifically for model m0.  The parameter array for this model is the empty array '[]'.
However the test for setting up the model-free models was being bypassed because the empty array
evaluates to False.  By removing the test for the parameter array and only checking for the model
name and equation type, the bug is removed.