Fix for bug #7297 (https://gna.org/bugs/index.php?7297).
This bug was reported by Alex Hansen.  The problem was that relax was being too strict with the
checking of the Da and Dr diffusion tensor parameters.  I have added a margin of error to these
checks to allow machine precision errors and optimisation precision issues (for the Method of
Multipliers constraint algorithm which allows the parameters outside of the limits).