Fixing the rest of bug #7241 (https://gna.org/bugs/?7241).
Bug #7241 was thought to be fixed in in r2591 and r2593, the commit messages describing the solution
being located at https://mail.gna.org/public/relax-commits/2006-09/msg00064.html (Message-id:
<E1GTgBi-0000R6-4h@subversion.gna.org>) for r2591 and
https://mail.gna.org/public/relax-commits/2006-10/msg00001.html (Message-id:
<E1GTt6C-0005rk-8q@subversion.gna.org>) for r2593.

However this was not the only place that the Scientific Python PDB data structure peptide_chains was
being accessed.  The chains were being accessed in the file 'generic_fns/sequence.py' when the
sequence was being read out of the PDB file.  This has now been modified with changes similar to
r2591 and r2593.