Complete overhaul of the relax manual targets.
The relax manual is now referred to as the user manaul (to allow a developers manual in the future).

The functions for compiling the manual have been renamed:
  'user_manual_html()' to 'compile_user_manual_html()'.
  'user_manual_pdf()' to 'compile_user_manual_pdf()'.

The header printed when compiling from the LaTeX sources has been updated.

Many target names have changes:
  'manual_dist' to 'user_manual_pdf_dist'.
  'manual' to 'user_manual_pdf'.
  'manual_nofetch' to 'user_manual_pdf_nofetch'.
  'manual_dist' to 'user_manual_pdf_dist'.
  'user_manual_pdf' to 'compile_user_manual_pdf'.
  'user_manual_html' to 'compile_user_manual_html'.

The HTML target has been split in two.