Added the API documentation creation (using Epydoc) to the Scons scripts.
The Scons target to create the HTML API documentation is called 'api_manual_html'.  The
documentation can be created by typing:
$ scons api_manual_html

The function 'compile_api_manual_html()' was added to the 'scons/manuals.py' file.  This function
runs the 'epydoc' command.  All the Epydoc options are specified in that function.