Fix for the documentation bug #7402 (https://gna.org/bugs/?7402).
This fixes the sigma(NOE) equations in section 8.8.1 of the relax user manual.  Alex Hansen reported
that the equations incorrect as one of the spectral density terms were multiplied by a factor of 6
when there should be nothing.  This problem does not affect the code, the equations there are
correct.