Fixes for a number of bugs caused by the changes for task #3838.
Task #3838 is located at https://gna.org/task/?3838 and more information is given at
https://mail.gna.org/public/relax-users/2006-10/msg00002.html (Message-id:
<7f080ed10610032153o4571935bq138d2739f55ca3bd@mail.gmail.com>).

Part of the problem, the failure of two of the tests of the test suite was reported by Alex Hansen
(viochemist _at_ gmail _dot_ com) in the post located at
https://mail.gna.org/public/relax-devel/2006-10/msg00095.html (Message-id:
<481156b20610130720u39a93e30k6d4603efd8730c96@mail.gmail.com>).  Another failure was reported by
Alex at https://mail.gna.org/public/relax-devel/2006-10/msg00096.html (Message-id:
<481156b20610130728jb6aa76ag3d9487a20397391b@mail.gmail.com).

The first fix, which was a missing run argument, is described at
https://mail.gna.org/public/relax-devel/2006-10/msg00099.html (Message-id:
<7f080ed10610130741j146721c8w512ba61bac3644c7@mail.gmail.com>).

The second problem was that the residue number and residue name was being referred to as 'num' and
'name' when they should have been 'res_num' and 'res_name'.

The third problem was an index for selecting the residue.  Rather than using the real index 'i', the
index 'j' was being used.