Created the user function 'pdb.create_tensor_pdb()'.
This is in response to the thread stating at
https://mail.gna.org/public/relax-devel/2006-10/msg00136.html (Message-id:
<7f080ed10610252045r3dcbba38q58aa53142af96e5a@mail.gmail.com>).

This new user function creates a PDB file of artificial structures to represent the diffusion
tensor.  The prompt interface function has been completed and the shell of the function
'self.create_tensor_pdb()' has been added to the file 'generic_fns/pdb.py'.  This second function
solely tests that the run exists, that a PDB structure has been loaded, that the sequence has been
loaded, and then opens and then closes the PDB file.  Hence an empty file is created by the user
function.