An overhaul of the optimisation tests of the relax test suite.
This is in response to the many posts in the thread starting at
https://mail.gna.org/public/relax-devel/2006-10/msg00114.html (Message-id:
<7f080ed10610190804w5681fafav843718f50f985f40@mail.gmail.com>).

The tests have been modified to check all parameter values and optimisation statistic prior to
failure.  All the faulty values will hence be printed out.  This idea is from the post at
https://mail.gna.org/public/relax-devel/2006-10/msg00128.html (Message-id:
<7f080ed10610200854v7d5288b6yc01d8c2ed0557abe@mail.gmail.com>).

The optimisation statistics (iteration count, function count, gradient count, and Hessian count)
tests have been changed.  The 'self.test_values()' function now accepts the 'max' flag argument.
The tests of these stats set the flag to 1 causing the test to fail only if the values are greater
than the expected value.  This allows for early termination of optimisation due to precision issues.
This idea is Gary's point number 2 in the post located at
https://mail.gna.org/public/relax-devel/2006-10/msg00117.html (Message-id:
<4538955F.3090300@bmb.leeds.ac.uk>).

Optimisation warnings are now tested for.  If the residue specific 'warning' data structure is not
None, the warning string is printed and the test fails.  This idea originated from Chris' post at
https://mail.gna.org/public/relax-devel/2006-10/msg00121.html (Message-id:
<1161343339.7703.46.camel@mrspell>).

The formula for the parameter value and chi-squared value test has been changed.  Previously the
test was:

    if val2 < val1 * (1 - error) or val2 > val1 * (1 + error):

Now the test is:

    if val2*scale < val1*scale - error or val2*scale > val1*scale + error:

This fixes the problem of testing chi-squared value.  Now the chi-squared value must be within 1e-20
of the expected value.  For the steepest descent which is terminated early, the value must be within
1e-8.