The diffusion tensor data structure now returns unit vectors parallel to the tensor axes.
The 'self.__getattr__()' function has been expanded.  The docstring now describes all the equations
used in the calculation of the parameters Diso, Dper, Dpar, Dratio, Dx, Dy, and Dz.  It also
describes how the unit vectors parallel to Dpar, Dx, Dy, and Dz are related to the spherical or
Euler angles.

Dper, Dpar, and Dratio are now only calculated if the diffusion tensor is spheroidal and Dx, Dy, and
Dz are only calculated if the diffusion is ellipsoidal.

The Dpar, Dx, Dy, and Dz unit vectors are calculated and a Numeric array is returned.