The length of the vector representing the unique axis of the spheroid is now proportional to Dpar.
To calculate the new vector, the Dpar unit vector is multiplied by the eigenvalue Dpar.  It is also
multiplied by the scaling factor 'scale'.  This scaling factor comes from the prompt interface user
function 'pdb.create_tensor_pdb()' which defaults to 1.8e-6.  The result is that the correlation
times translate to:

3 ns -> 100 Angstrom
10 ns -> 30 Angstrom
30 ns -> 10 Angstrom

The docstring of the 'pdb.create_tensor_pdb()' user function has been expanded to include a
description of the scaling factor and how the size of the tensor geometric object is larger for
smaller objects.