Created the geometric representation of the diffusion tensor in the tensor PDB file.
The lengths of the uniformly distributed vectors have been extended and scaled to the shape of the
geometric object.  Each atom has been connected to the previous atom in its radial array (generating
longitudinal lines) and the atoms of each radial array have been joined to the equivalent atoms of
the previous array (generating latitudinal lines).  To zip up a gap in the geometric object where
the latitudinal lines are not continuous ellipses, the atoms of the last radial array have been
joined to the first radial array.