Modification to the uniform distribution of vectors.
The vectors of the radial array (longitude) have been shifted by half an increment and the last
vector, which has moved beyond the pole, has been dropped.  In addition, vectors have been added to
the top and bottom poles, slightly breaking the uniformity of the distribution.  This improves
the appearance of the PDB geometric object representing the diffusion tensor.