Fix for bug #7616 (https://gna.org/bugs/index.php?7616).
This fixes the failure of the model selection between the global diffusion models when the local tm
models are included.  This is a very important bug fix.

The 'model_statistics()' function in the model-free and hybrid files have been modified.  The
'min_instances' argument to the function was not being utilised at all since r2516
(https://mail.gna.org/public/relax-commits/2006-07/msg00010.html) which introduced this bug.  The
argument has been renamed to 'global_stats'.  This 'global_stats' argument is feed in solely by the
model selection code.  That code decides if global statistics (which are summed for the local tm
models) should be utilised using the test:

    if self.num_instances[j] > self.min_instances or self.num_instances[j] == 1:

The model selection function, if the global stats argument is 1 and 'self.param_set' is equal to
'local_tm', will sum the chi2 values of the individual local tm models and return that value.

To fix an issue with the OpenDX code causing a failure of the test suite, the 'min_instances'
argument in the call to the model statistics function has been dropped.