Using the new 'self.atomic_data' data structure, the tensor PDB file is now created correctly.
The centre of mass atom, diffusion tensor representative atoms, and the atoms representing the axes
now belong to separate residues, COM, TNS, and AXS respectively.  A large number of changes were
necessary to fix the breakages introduced by the change in 'self.atomic_data'.

The atom number in 'self.atomic_data' was one out.

The atom name for the tensor atoms was being set to the 'atom_id' string.  This was incorrect as the
protons should have been named H1, H2, H3, etc.  The 'atom_id' string was too long for the HETATM
record.

In the 'self.write_pdb_file()' function, the elements of the 'het_data' structure used for
non-standard residues has been documented in the docstring.  This structure is now created prior to
the creation of the HET, HETNAM, and FORMUL records.

All the other records (HETATM, TER, CONECT, MASTER, END) are now correctly created.