'pymol.tensor_pdb()' has been modified to change the representation of 'COM', 'AXS', and 'SIM'.
The centre of mass 'COM' is shown as the 'dots' representation and is coloured blue.

The diffusion tensor axes 'AXS' are shown as the 'sticks' representation, is coloured cyan, and the
atoms labelled to show 'Dpar', 'Dx', 'Dy', and 'Dz' in PyMOL.

The Monte Carlo simulations axes are coloured cyan.