Added extra atoms to the 'AXS' residue of the tensor PDB representation for labelling.
Two or six extra atoms are added for the spheroid and ellipsoid axes respectively.  These are set to
be nitrogen atoms to allow easy selection of just those atoms.  They are also placed 3 Angstrom
outside of the geometric object so that the labels do not clash with the latitudinal and
longitudinal lines as well as the Monte Carlo simulation axes.

The code for generating the PDB chemical formula in 'self.write_pdb_file()' was modified to handle
nitrogen atoms.

The PyMOL user function 'pymol.tensor_pdb()' has been modified to use the new representation.  The
final selection is also deleted at the end of the user function.