The intensity reading function now counts the number of header lines in the Sparky or XEasy file.
The function 'self.number_of_header_lines()' has been added.  This attempts to execute the intensity
reading functions 'self.intensity_sparky()' or 'self.intensity_xeasy()' and if these fail with
either a RelaxError or an IndexError, then the line is considered a header line.  The function will
then return the number of header lines.