Added three flow diagrams to the relax manual describing the steps of model-free analysis.
The first diagram details the steps of the model-free optimisation protocol introduced by Mandel et
al., 1995.

The second presents the changes which occurred with the introduction of AIC model selection,
model elimination of failed models, and the use of all diffusion tensor types.

The last diagram presents the new model-free optimisation protocol encoded into the
'full_analysis.py' sample script (which is described in a manuscript which has been submitted for
publication).