Partial fix for bug #7787 (https://gna.org/bugs/?7787).
This bug was reported by Stephen Headey (https://gna.org/users/sjheadey).

The issue is that if the 'full_analysis.py' script is improperly used, the error messages don't tell
you what the real issue is.  In the bug report, the problem was that only a single round of
optimisation of the sphere, spheroids, and ellipsoid was being run.  The 'full_analysis.py' script
has therefore been modified to catch this and throw a RelaxError describing the issue.