Fix for bug #7755 (https://gna.org/bugs/?7755).
This bug was reported by Stephen Headey (https://gna.org/users/sjheadey).

The problem was that the select_sim array created by the columnar results reading function
'self.read_columnar_results()' in 'specific_fns/model_free.py' was one massive single array
containing all the selected simulation data of all spin systems of all simulations.  Subsequently
this was then not being split up for each spin system.  Hence the total number of simulations was
being set to the number of spin systems times the number of simulations.  The select_sim array for
each spin system was also of the same length and contained the data for all spins.

The fix was to convert the 'select_sim' array being created by 'self.read_columnar_results()' into
a Numeric matrix.  The first dimension corresponds to the simulation and the second corresponds to
the instance.