Added the method 'self.create_vector_dist()' to 'generic_fns/structure.py'.
This function currently calculates and adds the centre of mass to the PDB file, adds the XH vectors
as protons to the 'self.atomic_data' data structure, and creates the PDB file.  The code is
incomplete as each vector is being set to the true residue name and number and ATM record support
still needs to be added.