Ported r2924 to r2927 from the 1.3 line.
The command used was:
svn merge -r2923:2927 svn+ssh://bugman@svn.gna.org/svn/relax/1.3

This adds a few features to the selection user functions.  For the the 'select.read()' and
'unselect.read()' user functions a different column for the residue numbers can now be specified.
The 'select.res()' and 'select.read()'  now accept the boolean keyword option which can be set to
'OR', 'NOR', 'AND', 'NAND', 'XOR', or 'XNOR'.  The default, and previous behaviour, is 'OR'.