Ported r2931 from the 1.3 line.
The command used was:
svn merge -r2930:2931 svn+ssh://bugman@svn.gna.org/svn/relax/1.3

The commit message of the revision is:

Bug fix for the plotting of model-free results within Molmol macros.

The fast correlation time macro (te < 100 ps and tf) was setting bond lengths to zero when there
was no parameter.  This has been changed so that the bond is now set to a width of 0.3 and is
coloured white.